Extract out the Tp.BaseDevice compat code into a new file
Makes it clear that this code exists only for compatibility
with the current version of the thingpedia-api library (and a couple
of Thingpedia devices themselves) and might go away soon
