+++ lib/loaders/base_generic.js
@@ -11,6 +11,7 @@
 
 const Tp = require('thingpedia');
 const Utils = require('../utils');
+const { makeBaseDeviceMetadata } = require('../compat');
 const ConfigMixins = require('../config');
 
 module.exports = class BaseGenericModule {
@@ -51,7 +52,7 @@ module.exports = class BaseGenericModule {
         };
         this._config.install(this._loaded);
 
-        this._loaded.metadata = Utils.makeBaseDeviceMetadata(this._manifest);
+        this._loaded.metadata = makeBaseDeviceMetadata(this._manifest);
     }
 
     get id() {
