+++ lib/loaders/base_js.js
@@ -14,7 +14,8 @@ const stream = require('stream');
 
 const Tp = require('thingpedia');
 
-const { makeBaseDeviceMetadata, getPollInterval } = require('../utils');
+const { getPollInterval } = require('../utils');
+const { makeBaseDeviceMetadata } = require('../compat');
 const { ImplementationError } = require('../errors');
 const ConfigMixins = require('../config');
