+++ lib/loaders/unsupported_builtin.js
@@ -13,6 +13,7 @@ const Tp = require('thingpedia');
 const ThingTalk = require('thingtalk');
 
 const Utils = require('../utils');
+const { makeBaseDeviceMetadata } = require('../compat');
 const { UnsupportedError } = require('../errors');
 
 module.exports = class UnsupportedBuiltinModule {
@@ -57,7 +58,7 @@ module.exports = class UnsupportedBuiltinModule {
             };
         }
 
-        this._loaded.metadata = Utils.makeBaseDeviceMetadata(manifest);
+        this._loaded.metadata = makeBaseDeviceMetadata(manifest);
     }
 
     get id() {
